initial restructure into separate files
- preparing for eventual packaging
- most docs are going to need updating now
    - e.g. updating links to :func:`thing` etc
- minimal testing of restructure performed
- parsing of doxygen manually next
