+++ .flake8
@@ -1,0 +1,16 @@
+[flake8]
+max-line-length = 111
+ignore =
+    # I like to line up operators when possible
+    E221, E222,
+    # I also like obscene block comments for the minimap on sublime
+    E266,
+    # And sometimes I likes to line up my parameters
+    E241,
+    # And sometimes I likes to line up my keywords
+    E272
+exclude =
+    # Do not parse the git directory
+    .git,
+    # Ignore the docs conf.py from sphinx
+    docs/conf.py,
